feat: improve gke setup and tear down

+++ k8s/tear_down_gcloud_project.sh
@@ -1,5 +1,10 @@
 #!/bin/bash -ex
 
+if [[ "$1" = "" ]]; then
+  echo "usage specify PROJECT_ID";
+  exit 1;
+fi
+
 PROJECT_ID=test-gitlabci-k8s
 
 gcloud projects delete $PROJECT_ID
